Handle failing transports and report failures to the client
We don't tell the clients which transports failed for privacy reasons.
They only get to know if none, some or all of them failed.

In the future, we may implement retrying to send with the failed
transports using the message id but this means that we would have to
store which transports failed.